Factored out Scalene magic numbers into scalene_config; fixed viewer loading.
