+++ scalene/scalene_parseargs.py
@@ -7,7 +7,7 @@ from typing import Any, List, NoReturn, Optional, Tuple
 
 from scalene.find_browser import find_browser
 from scalene.scalene_arguments import ScaleneArguments
-from scalene.scalene_version import scalene_version, scalene_date
+from scalene.scalene_config import scalene_version, scalene_date
 
 scalene_gui_url = f'file:{os.path.join(os.path.dirname(__file__), "scalene-gui", "index.html")}'
 
@@ -365,13 +365,14 @@ for the process ID that Scalene reports. For example:
             if browser := find_browser():
                 assert not args.no_browser
                 dir = os.path.dirname(__file__)
+                import scalene.scalene_config
                 import subprocess
                 subprocess.Popen([sys.executable,
                                   f"{dir}{os.sep}launchbrowser.py",
                                   "demo",
-                                  "11235"]) # FOR NOW WAS str(SCALENE_PORT)])
-    #                             stdout=subprocess.DEVNULL,
-    #                             stderr=subprocess.DEVNULL)
+                                  str(scalene.scalene_config.SCALENE_PORT)],
+                                 stdout=subprocess.DEVNULL,
+                                 stderr=subprocess.DEVNULL)
                 sys.exit(0)
                 pass
             else:
