--- scalene/scalene_version.py
@@ -1,4 +1,0 @@
-"""Current version of Scalene; reported by --version."""
-
-scalene_version = "1.5.34"
-scalene_date = "2024.01.27"
